REMB: Do not pass the len to the packet constructor (not needed)

+++ worker/src/RTC/WebRtcTransport.cpp
@@ -1269,8 +1269,7 @@ namespace RTC
 			  ssrcsStream.str().c_str());
 		}
 
-		// Avoid ambiguous constructor.
-		RTC::RTCP::FeedbackPsRembPacket packet(static_cast<uint32_t>(0), 0);
+		RTC::RTCP::FeedbackPsRembPacket packet(0, 0);
 
 		packet.SetBitrate(effectiveBitrate);
 		packet.SetSsrcs(ssrcs);
